+++ doc/releaseNotes.md
@@ -511,3 +511,9 @@ This file describes the main feature changes for each readout.exe released versi
 - Updated configuration parameters:
   - added consumer-fileRecorder-*.dropEmptyHBFramesTriggerMask: when using dropEmptyHBFrames = 1, keep some empty HB frames with trigger type matching given (decimal) mask.
 - Added support for RDHv7. Readout already accepts data from equipments with v7, but still generates data (internal sw generator) with v6, for compatibility with current software downstream.
+
+## next version
+- Updated configuration parameters:
+  - added equipemnt-*.rdhCheckTrigger: when set, enables some consistency checks on trigger counters, as follows:
+    - Trigger type TF should happen only on orbits % TF period = 0.
+    - First trigger received on each link should be SOC or SOT.